requirements.txt
numpy
scipy
pyyaml
tqdm
matplotlib
opencv-python
torch
torchvision
ultralytics
